WIP: getting the rest of the Stripe payment process in place, undoing some reorg I tried to do
