+++ src/payment_gateway/mitol/payment_gateway/constants.py
@@ -6,6 +6,10 @@ MITOL_PAYMENT_GATEWAY_NONE = "None"
 MITOL_PAYMENT_GATEWAY_CYBERSOURCE = "CyberSource"
 MITOL_PAYMENT_GATEWAY_STRIPE = "Stripe"
 
+CART_ITEM_INLINE = "inline"
+CART_ITEM_DEFINED = "defined"
+CART_ITEM_UNKNOWN = "unknown"
+
 CYBERSOURCE_DECISION_ACCEPT = "ACCEPT"
 CYBERSOURCE_DECISION_DECLINE = "DECLINE"
 CYBERSOURCE_DECISION_REVIEW = "REVIEW"
